Actualización de sistema web

+++ Sistema web/app/models/measure.rb
@@ -7,9 +7,9 @@ class Measure < ApplicationRecord
   def decrypted_value
     # The value should be in base 64
     # If value is plain encrypted append { format: :plain }
-    # AES.decrypt(self[:value], "una contraseña")
+    # AES.decrypt(self[:value], 'aaaabbbbccccdddd', { format: :plain })
 
     # Decode base 64
-    Base64.decode64(self[:value]).unpack("A*").first.to_f
+    Base64.decode64(self[:value]).unpack("A*").first.to_i
   end
 end
